Start adding a basic part addition ui
